+++ lib/async.dart
@@ -1,5 +1,6 @@
 /// Provides utilities for asynchronous programming, such as `ValueSubject`.
 library async;
 
-export './src/async/value_subject.dart';
 export './src/async/stream_combine_latest.dart';
+export './src/async/stream_merge.dart';
+export './src/async/value_subject.dart';